+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/InlineQueries/ChosenInlineResult/ChosenInlineResult.kt
@@ -1,10 +1,10 @@
 package dev.inmo.tgbotapi.types.InlineQueries.ChosenInlineResult
 
+import dev.inmo.tgbotapi.CommonAbstracts.FromUser
 import dev.inmo.tgbotapi.types.*
 
-sealed interface ChosenInlineResult {
+sealed interface ChosenInlineResult : FromUser {
     val resultId: InlineQueryIdentifier //chosen temporary, can be changed
-    val user: User
     val inlineMessageId: InlineMessageIdentifier?
     val query: String
 }