+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/InlineQueries/ChosenInlineResult/RawChosenInlineResult.kt
@@ -19,7 +19,7 @@ internal data class RawChosenInlineResult(
     val inlineMessageId: InlineMessageIdentifier? = null
 ) {
     val asChosenInlineResult: ChosenInlineResult by lazy {
-        location?.let {
+        location ?.let {
             LocationChosenInlineResult(resultId, user, location, inlineMessageId, query)
         } ?: BaseChosenInlineResult(resultId, user, inlineMessageId, query)
     }